requirements.txt
pandas>=2.2.2
pyarrow>=16.1.0
python-dotenv>=1.0.1
azure-storage-blob>=12.20.0